Create logger.py
This script logs the startup information and the contents of the startup directory to a log file located at 'startup_log.txt'. The logging level is set to 'INFO', so only messages with a severity level of 'INFO' or higher will be logged. You can customize the logging level and format to meet your specific needs.

Note that this script will only log the startup information for the current user. If you want to log the startup information for all users, you will need to modify the script accordingly. Additionally, running a script to log everything when you turn on your PC may have security and privacy implications, so use this script responsibly and with caution.